Tests/LibWeb/Text/input/css/getComputedStyle-item-does-not-return-invalid-propertyid-on-length-boundary.html

<!DOCTYPE html>
<script src="../include.js"></script>
<script>
    test(() => {
        const style = getComputedStyle(document.body);
        println(`getComputedStyle().item(length) should return empty string: '${style.item(style.length)}'`);
    });
</script>
