Tests/LibWeb/Text/expected/css/getComputedStyle-item-does-not-return-invalid-propertyid-on-length-boundary.txt

getComputedStyle().item(length) should return empty string: ''
